+++ src/NaturalResultType.php
@@ -16,6 +16,9 @@ abstract class NaturalResultType
 
     const KNOWLEDGE = 'knowledge';
 
+    const PEOPLE_ALSO_ASK = 'people_also_ask';
+    const PAA_QUESTION = 'paa_question';
+
     const IMAGE_GROUP = 'image_group';
     const IMAGE_GROUP_IMAGE = 'image_group_image';
 
@@ -24,6 +27,9 @@ abstract class NaturalResultType
 
     const IN_THE_NEWS = 'in_the_news';
     const TOP_STORIES = 'top_stories';
+    const TOP_STORIES_NEWS_VERTICAL = 'top_stories_news_vertical';
+    const TOP_STORIES_NEWS_CAROUSEL = 'top_stories_news_carousel';
+    const TOP_STORIES_COMPOSED = 'top_stories_composed';
     const TWEETS_CAROUSEL= 'tweets_carousel';
 
     const MAP = 'map';