+++ src/Parser/Evaluated/MobileNaturalParser.php
@@ -7,20 +7,21 @@ namespace Serps\SearchEngine\Google\Parser\Evaluated;
 
 use Serps\SearchEngine\Google\Page\GoogleDom;
 use Serps\SearchEngine\Google\Parser\AbstractParser;
-use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\AnswerBox;
-use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalResult;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalCardsResultO9g5cc;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalCardsResultZ1m;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalCardsResultZINbbc;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalCardsVideoResult;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\ComposedTopStories;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Divider;
-use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Flight;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\ImageGroup;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\ImageGroupCarousel;
-use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\InTheNews;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\LargeClassicalResult;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\KnowledgeCard;
-use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Map;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalCardsResult;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\MapMobile;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\PeopleAlsoAsk;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\SearchResultGroup;
-use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\TweetsCarousel;
-use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalWithLargeVideo;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\TweetsCarouselZ1m;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\VideoGroup;
 
 /**
@@ -37,10 +38,18 @@ class MobileNaturalParser extends AbstractParser
         return [
             new Divider(),
             new SearchResultGroup(),
+            new ClassicalCardsResultO9g5cc(),
+            new ClassicalCardsResultZINbbc(), // TODO maybe outdated
+            new ClassicalCardsResultZ1m(), // TODO remove (outdated)
+            new ClassicalCardsVideoResult(),
             new ClassicalCardsResult(),
+            new MapMobile(),
+            new TweetsCarouselZ1m(), // TODO replace and remove (outdated)
             new ImageGroupCarousel(),
+            new ComposedTopStories(),
             new VideoGroup(),
             new ImageGroup(),
+            new PeopleAlsoAsk(), // people also ask must be placed before knowledge card to stop parsing
             new KnowledgeCard()
         ];
     }