+++ src/Parser/Evaluated/Rule/Natural/AnswerBox.php
@@ -17,7 +17,10 @@ class AnswerBox implements ParsingRuleInterface
     public function match(GoogleDom $dom, \Serps\Core\Dom\DomElement $node)
     {
         if ($node->getAttribute('class') == 'g mnr-c g-blk'
-            && $dom->cssQuery('._Z7', $node)->length == 1
+            && (
+                $dom->cssQuery('.ifM9O', $node)->length == 1 ||
+                $dom->cssQuery('._Z7', $node)->length == 1      // TODO used for BC, remove in the future
+            )
         ) {
             return self::RULE_MATCH_MATCHED;
         }
@@ -55,7 +58,8 @@ class AnswerBox implements ParsingRuleInterface
                 return $citeTag->nodeValue;
             },
             'description' => function () use ($dom, $node) {
-                $citeTag = $dom->cssQuery('.mod ._Tgc', $node)
+                // TODO "mod ._Tgc" kept for BC, remove in the future
+                $citeTag = $dom->cssQuery('.mod ._Tgc, .mod .Y0NH2b', $node)
                     ->item(0);
                 if (!$citeTag) {
                     // TODO ERROR