+++ test/suites/GoogleUrlTest.php
@@ -20,7 +20,7 @@ class GoogleUrlTest extends \PHPUnit_Framework_TestCase
     public function testConstruct()
     {
         $googleUrl = new GoogleUrl();
-        $this->assertEquals('google.com', $googleUrl->getHost());
+        $this->assertEquals('www.google.com', $googleUrl->getHost());
         $this->assertEquals('/search', $googleUrl->getPath());
     }