+++ CHANGELOG.md
@@ -1,12 +1,133 @@
 # CHANGELOG
 
+## 0.4.7
+
+*2018-18-05*
+
+* Bug fix: 
+    * Fixed title for adwords results
+
+## 0.4.6
+
+*2018-30-10*
+
+* Feature: 
+    * Added adwords results for mobiles
+    * Added map results for mobiles
+* Bug fix:
+    * Fixed description for mobile classical
+    * Fixed related searches for mobile
+* Other:
+    * Added ``ext-dom`` in composer.json
+
+
+## 0.4.5
+
+*2018-30-10*
+
+* Bug fix:
+    * Fixed local pack (#113) Thanks @Human018
+    * Fixed urls for ads results
+
+## 0.4.4
+
+*2018-10-22*
+
+* Bug fix:
+    * Fixed people also ask update
+
+## 0.4.3
+
+*2018-09-17*
+
+* Bug fix:
+    * Fix google dom update on classical results
+
+## 0.4.2
+
+*2018-08-05*
+
+* Bug fix:
+    * fix mobile serps (#106)
+
+
+## 0.4.1
+
+*2018-07-05* 
+
+* Bug fix:
+    * fixed multiple mobile issues on mobile results
+    * fixed parsing for number of results (#100) - thanks @migliori
+    * fixed related searches on desktop  - thanks @gudix
+
+## 0.4.0
+
+*2018-05-29*
+
+* Bug fix:
+    * fixed the captcha exception. The right exception is now returned when a captcha is found
+    * fixed invalid type hinting causing errors with hhvm
+* Google updates:
+    * **bc break** removed support for image captcha as google now uses recaptcha
+* Other:
+    * When an invalid classical result is found, throw an exception instead of returning invalid results causing fatal errors.
+
 ## 0.3.0 
 
-*20xx-xx-xx*
+*2018-04-04*
+
+* Dependencies
+    * **bc break** use version ``0.3.x`` of ``serps/core``
+* Updates
+    * **bc break** google url default domain is now ``"www.google.com`` instead of ``google.com``. This way we avoid extra redirects too the ``"www"`` subdomain.
+    * Fix a bug with search result group parser that was triggering a php error.
+* Dom Updates
+    * Fix parsing for classical results on mobiles.
+    * Fix parsing for knowledge cards on mobiles.
+
+## 0.2.5
+
+*2018-03-29*
+
+* Bug fix:
+    * Fix a bug with map results introduced in version 0.2.4 see [#94](https://github.com/serp-spider/search-engine-google/issues/94)
+
+## 0.2.4
+
+*2018-03-22*
+
+* Bug fix:
+    * Fix google update for map results
+    * Fix google update for "destination" data in classical results
+    * Fix google update for People Also Ask
+    * Fix google update for answer box [#90](https://github.com/serp-spider/search-engine-google/issues/90)
+    
+## 0.2.3
+
+*2017-12-11*
+
+* Features:
+    * Added parsing for people also ask results [#70](https://github.com/serp-spider/search-engine-google/issues/70)
+
+* Bug fix:
+    * Fix some mobile card results not parsing [#83](https://github.com/serp-spider/search-engine-google/issues/83)
 
 ## 0.2.2
 
-*20xx-xx-xx*
+*2017-11-25*
+
+* Bug fix:
+    * Parse ``bkWMgd`` groups (thanks to [Shiftas](https://github.com/Shiftas)) [#76](https://github.com/serp-spider/search-engine-google/issues/76)
+    * Fix result count [#76](https://github.com/serp-spider/search-engine-google/issues/76)
+    * Fix some mobile card results not parsing [#79](https://github.com/serp-spider/search-engine-google/issues/79) and [#78](https://github.com/serp-spider/search-engine-google/issues/78)
+    * Fix twitter carousel parser for mobile [#81](https://github.com/serp-spider/search-engine-google/issues/81)
+    * Fix related searches for mobile [#80](https://github.com/serp-spider/search-engine-google/issues/81)
+
+* Features:
+    * Parsing for "composed top stories" and standardizing old "top stories" [#67](https://github.com/serp-spider/search-engine-google/issues/67)
+
+* Other:
+    * Dependency to serps/core was updated from ~0.2.0 to ~0.2.4
 
 ## 0.2.1